refactor(订单): 将金额类型从float64改为int以分单位存储
修改订单相关结构体和服务层代码，使用整数分单位代替浮点数元单位
添加dmoney包处理金额转换逻辑
更新数据库表名和字段映射

+++ internal/interface/handler/order_handler.go
@@ -26,8 +26,8 @@ func (h *OrderHandler) CreateOrder(w http.ResponseWriter, r *http.Request) {
 		Items      []struct {
 			ProductID string  `json:"product_id"`
 			Quantity  int     `json:"quantity"`
-			UnitPrice float64 `json:"unit_price"`
-			Subtotal  float64 `json:"subtotal"`
+			UnitPrice int `json:"unit_price"`
+			Subtotal  int `json:"subtotal"`
 		} `json:"items"`
 	}
